Rename and delete old files
