+++ src/components/page-application/style.css
