--- src/components/page-application/styles.css
@@ -1,9 +1,0 @@
-.page-application > .app-summary {
-  margin-bottom: 1rem;
-}
-
-@media (max-width: 50rem) {
-  .page-application__content-header {
-    padding-left: 3.5rem;
-  }
-}
